+++ docs/changelog.md
@@ -263,6 +263,7 @@
 - simplify methods impls: `take_ascii_ident`, `take_ascii_ident_tail`.
 
 ### text::str
+- make `StringU*` use `NonMax*` for the length.
 - remove `StringU32`, `StringUSize`.
 
 ### ui::event